refactor export service to support more entities
