+++ src/main/java/org/candlepin/subscriptions/capacity/CapacityIngressConfiguration.java
@@ -21,8 +21,8 @@
 package org.candlepin.subscriptions.capacity;
 
 import org.candlepin.subscriptions.db.RhsmSubscriptionsDataSourceConfiguration;
+import org.candlepin.subscriptions.export.ExportSubscriptionConfiguration;
 import org.candlepin.subscriptions.resteasy.ResteasyConfiguration;
-import org.candlepin.subscriptions.subscription.export.ExportSubscriptionConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigurationExcludeFilter;
 import org.springframework.boot.context.TypeExcludeFilter;
 import org.springframework.context.annotation.ComponentScan;
@@ -46,7 +46,12 @@ import org.springframework.scheduling.annotation.EnableAsync;
   ExportSubscriptionConfiguration.class
 })
 @ComponentScan(
-    basePackages = {"org.candlepin.subscriptions.capacity", "org.candlepin.subscriptions.product"},
+    basePackages = {
+      "org.candlepin.subscriptions.capacity",
+      "org.candlepin.subscriptions.product",
+      "org.candlepin.subscriptions.export",
+      "org.candlepin.subscriptions.subscription.export",
+    },
     // Prevent TestConfiguration annotated classes from being picked up by ComponentScan
     excludeFilters = {
       @ComponentScan.Filter(type = FilterType.CUSTOM, classes = TypeExcludeFilter.class),
