+++ src/main/java/org/candlepin/subscriptions/export/ExportSubscriptionListener.java
@@ -18,9 +18,9 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.subscription.export;
+package org.candlepin.subscriptions.export;
 
-import static org.candlepin.subscriptions.subscription.export.ExportSubscriptionConfiguration.SUBSCRIPTION_EXPORT_QUALIFIER;
+import static org.candlepin.subscriptions.export.ExportSubscriptionConfiguration.SUBSCRIPTION_EXPORT_QUALIFIER;
 
 import com.fasterxml.jackson.core.JsonEncoding;
 import com.fasterxml.jackson.core.JsonGenerator;
@@ -36,36 +36,23 @@ import jakarta.ws.rs.core.Response.Status;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
-import org.candlepin.clock.ApplicationClock;
-import org.candlepin.subscriptions.db.SubscriptionRepository;
-import org.candlepin.subscriptions.db.model.DbReportCriteria;
-import org.candlepin.subscriptions.db.model.ServiceLevel;
-import org.candlepin.subscriptions.db.model.Subscription;
-import org.candlepin.subscriptions.db.model.Usage;
 import org.candlepin.subscriptions.exception.ExportServiceException;
-import org.candlepin.subscriptions.json.SubscriptionsExportItem;
-import org.candlepin.subscriptions.json.SubscriptionsExportMeasurement;
 import org.candlepin.subscriptions.rbac.RbacApiException;
 import org.candlepin.subscriptions.rbac.RbacService;
 import org.candlepin.subscriptions.task.TaskQueueProperties;
 import org.candlepin.subscriptions.util.KafkaConsumerRegistry;
 import org.candlepin.subscriptions.util.SeekableKafkaConsumer;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Profile;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
 
 /** Listener for Export messages from Kafka */
 @Service
 @Slf4j
-@Profile("capacity-ingress")
 public class ExportSubscriptionListener extends SeekableKafkaConsumer {
 
   public static final String ADMIN_ROLE = ":*:*";
@@ -75,26 +62,23 @@ public class ExportSubscriptionListener extends SeekableKafkaConsumer {
   private final ExportApi exportApi;
   private final ConsoleCloudEventParser parser;
   private final RbacService rbacService;
-  private final SubscriptionRepository subscriptionRepository;
   private final ObjectMapper objectMapper;
-  private final ApplicationClock clock;
+  private final List<DataExporterService<?, ?>> exporterServices;
 
   protected ExportSubscriptionListener(
       @Qualifier(SUBSCRIPTION_EXPORT_QUALIFIER) TaskQueueProperties taskQueueProperties,
-      SubscriptionRepository subscriptionRepository,
       KafkaConsumerRegistry kafkaConsumerRegistry,
       ConsoleCloudEventParser parser,
       ExportApi exportApi,
       RbacService rbacService,
       ObjectMapper objectMapper,
-      ApplicationClock clock) {
+      List<DataExporterService<?, ?>> exporterServices) {
     super(taskQueueProperties, kafkaConsumerRegistry);
-    this.subscriptionRepository = subscriptionRepository;
     this.parser = parser;
     this.exportApi = exportApi;
     this.rbacService = rbacService;
     this.objectMapper = objectMapper;
-    this.clock = clock;
+    this.exporterServices = exporterServices;
   }
 
   @Timed("rhsm-subscriptions.exports.upload")
@@ -112,14 +96,24 @@ public class ExportSubscriptionListener extends SeekableKafkaConsumer {
             Status.INTERNAL_SERVER_ERROR.getStatusCode(),
             "Cloud event doesn't have any Export data: " + request.getId());
       }
-      if (request.isRequestFor(SWATCH_APP)) {
-        log.info(
-            "Processing event: '{}' from application: '{}'", request.getId(), request.getSource());
-        checkRbac(request);
-        uploadData(request);
+      if (request.isRequestForApplication(SWATCH_APP)) {
+        exporterServices.stream()
+            .filter(s -> s.handles(request))
+            .findFirst()
+            .ifPresent(
+                exporterService -> {
+                  log.info(
+                      "Processing event: '{}' from application: '{}'",
+                      request.getId(),
+                      request.getSource());
+                  checkRbac(request);
+                  uploadData(exporterService, request);
 
-        log.info(
-            "Event processed: '{}' from application: '{}'", request.getId(), request.getSource());
+                  log.info(
+                      "Event processed: '{}' from application: '{}'",
+                      request.getId(),
+                      request.getSource());
+                });
       }
     } catch (ExportServiceException e) {
       log.error(
@@ -134,13 +128,13 @@ public class ExportSubscriptionListener extends SeekableKafkaConsumer {
     }
   }
 
-  private void uploadData(ExportServiceRequest request) throws ApiException {
+  private void uploadData(DataExporterService<?, ?> exporterService, ExportServiceRequest request) {
     // Here we need to determine format (csv or json)
-    var subscriptions = fetchData(request);
+    Stream<?> data = exporterService.fetchData(request);
     if (Objects.equals(request.getFormat(), Format.CSV)) {
       uploadCsv(request);
     } else if (Objects.equals(request.getFormat(), Format.JSON)) {
-      uploadJson(subscriptions, request);
+      uploadJson(exporterService, data, request);
     } else {
       throw new ExportServiceException(Status.FORBIDDEN.getStatusCode(), "Format isn't supported");
     }
@@ -152,55 +146,35 @@ public class ExportSubscriptionListener extends SeekableKafkaConsumer {
         Status.NOT_IMPLEMENTED.getStatusCode(), "Export upload csv isn't implemented ");
   }
 
-  private void uploadJson(Stream<Subscription> subscriptions, ExportServiceRequest request) {
+  private void uploadJson(
+      DataExporterService<?, ?> exporterService, Stream<?> data, ExportServiceRequest request) {
     log.debug("Uploading Json for request {}", request.getExportRequestUUID());
     File file = createTemporalFile(request);
-    try (subscriptions;
+    try (data;
         FileOutputStream stream = new FileOutputStream(file);
         JsonGenerator jGenerator = objectMapper.createGenerator(stream, JsonEncoding.UTF8)) {
       jGenerator.writeStartObject();
       jGenerator.writeArrayFieldStart("data");
       var serializerProvider = objectMapper.getSerializerProviderInstance();
       var serializer =
-          serializerProvider.findTypedValueSerializer(SubscriptionsExportItem.class, false, null);
-      subscriptions.forEach(
-          subscription -> {
-            var item = new SubscriptionsExportItem();
-            item.setOrgId(subscription.getOrgId());
-            item.setMeasurements(new ArrayList<>());
-
-            // map offering
-            var offering = subscription.getOffering();
-            item.setSku(offering.getSku());
-            Optional.ofNullable(offering.getUsage()).map(Usage::getValue).ifPresent(item::setUsage);
-            Optional.ofNullable(offering.getServiceLevel())
-                .map(ServiceLevel::getValue)
-                .ifPresent(item::setServiceLevel);
-            item.setProductName(offering.getProductName());
-            item.setSubscriptionNumber(subscription.getSubscriptionNumber());
-            item.setQuantity(subscription.getQuantity());
-
-            // map measurements
-            for (var entry : subscription.getSubscriptionMeasurements().entrySet()) {
-              var measurement = new SubscriptionsExportMeasurement();
-              measurement.setMeasurementType(entry.getKey().getMeasurementType());
-              measurement.setCapacity(entry.getValue());
-              measurement.setMetricId(entry.getKey().getMetricId());
-
-              item.getMeasurements().add(measurement);
-            }
-            try {
-              serializer.serialize(item, jGenerator, serializerProvider);
-            } catch (IOException e) {
-              log.error(
-                  "Error serializing the subscription {} for request {}",
-                  subscription,
-                  request.getExportRequestUUID(),
-                  e);
-              throw new ExportServiceException(
-                  Status.INTERNAL_SERVER_ERROR.getStatusCode(), "Error writing the Json payload");
-            }
-          });
+          serializerProvider.findTypedValueSerializer(
+              exporterService.getExportItemClass(), false, null);
+      data.map(exporterService::mapDataItem)
+          .forEach(
+              item -> {
+                try {
+                  serializer.serialize(item, jGenerator, serializerProvider);
+                } catch (IOException e) {
+                  log.error(
+                      "Error serializing the data item {} for request {}",
+                      item,
+                      request.getExportRequestUUID(),
+                      e);
+                  throw new ExportServiceException(
+                      Status.INTERNAL_SERVER_ERROR.getStatusCode(),
+                      "Error writing the Json payload");
+                }
+              });
       jGenerator.writeEndArray();
       jGenerator.writeEndObject();
     } catch (IOException e) {
@@ -239,42 +213,6 @@ public class ExportSubscriptionListener extends SeekableKafkaConsumer {
       return;
     }
     throw new ExportServiceException(Status.FORBIDDEN.getStatusCode(), "Insufficient permission");
-  }
-
-  private Stream<Subscription> fetchData(ExportServiceRequest request) throws ApiException {
-    // Check subscription table for data for the event
-    log.debug("Fetching subscriptionOrg for {}", request.getOrgId());
-    var reportCriteria = extractExportFilter(request);
-    return subscriptionRepository.streamBy(reportCriteria);
-  }
-
-  private DbReportCriteria extractExportFilter(ExportServiceRequest request) {
-    var report =
-        DbReportCriteria.builder()
-            .orgId(request.getOrgId())
-            .beginning(clock.now())
-            .ending(clock.now());
-    if (request.getFilters() != null) {
-      var filters = request.getFilters().entrySet();
-      try {
-        for (var entry : filters) {
-          var filterHandler =
-              ExportSubscriptionRequestFilters.get().get(entry.getKey().toLowerCase(Locale.ROOT));
-          if (filterHandler == null) {
-            log.warn("Filter '{}' isn't currently supported. Ignoring.", entry.getKey());
-          } else if (entry.getValue() != null) {
-            filterHandler.accept(report, entry.getValue().toString());
-          }
-        }
-
-      } catch (IllegalArgumentException ex) {
-        throw new ExportServiceException(
-            Status.BAD_REQUEST.getStatusCode(),
-            "Wrong filter in export request: " + ex.getMessage());
-      }
-    }
-
-    return report.build();
   }
 
   private static File createTemporalFile(ExportServiceRequest request) {
